test_requirements.txt
pytest==7.4.0
pytest-mock==3.12.0
requests-mock==1.11.0
orodha-keycloak==1.2.0
